Code: Compiler warnings fix
Example 2 was causing some compile warnings related to casting: using 1.0f instead of 1 now.

+++ examples/example2_Exponentially_weighted_moving_average.cpp
@@ -40,7 +40,7 @@ int main()
         constexpr float noise_std_dev = 1.5;
 
         
-        std::iota(timeData.begin(),timeData.end(),1); // Fill time with 1,2,3,...
+        std::iota(timeData.begin(),timeData.end(),1.0f); // Fill time with 1,2,3,...
         std::transform(timeData.begin(),timeData.end(),inputData.begin(),[&](float time){   // Compute input
             return SineWaveWithNoiseSample(time,frequency,gain,noise_std_dev);
         });
@@ -84,6 +84,6 @@ int main()
 float SineWaveWithNoiseSample(const float time,const float frequency, const float gain, const float noise_std_dev ){
     static std::default_random_engine generator;
     static std::normal_distribution<float> distribution(0,noise_std_dev); // mean,sdev
-    return std::sin(2*PI*time/frequency)*gain + distribution(generator);
+    return (float) std::sin(2.0*PI*time/frequency)*gain + distribution(generator);
 }
 /// @endcond
